Fix counting frames in trace file output.
Fix integer overflow when frame skip count is non-zero and frame count is zero.

+++ VkLayer_profiler_layer/profiler_trace/profiler_trace.h
@@ -155,7 +155,7 @@ namespace Profiler
 
         std::string m_OutputFileName;
 
-        uint64_t m_MaxFrameCount;
+        uint32_t m_MaxFrameCount;
         uint32_t m_SkipFrameCount;
         std::atomic_uint32_t m_ProcessedFrameCount;
         std::atomic_bool m_Flushed;
